fix all imports for fitting.py
