+++ src/qicklab/utils/ana_utils.py
@@ -107,3 +107,32 @@ def get_longest_continuous_segment(segments_time, segments_vals):
     max_idx = np.argmax(lengths)
     return segments_time[max_idx], segments_vals[max_idx]
 
+def max_offset_difference_with_x(x_values, y_values, offset):
+    """
+    Find the x-value corresponding to the maximum average absolute difference from a given offset.
+
+    The function calculates the average absolute difference from the offset for triplets of consecutive y-values
+    and returns the x-value (from the middle point of each triplet) that corresponds to the maximum difference.
+
+    Parameters:
+        x_values (array_like): Array of x-values.
+        y_values (array_like): Array of y-values.
+        offset (float): The offset value to compare against.
+
+    Returns:
+        tuple:
+            corresponding_x: The x-value corresponding to the maximum average difference.
+            max_average_difference: The maximum average absolute difference calculated.
+    """
+    max_average_difference = -1
+    corresponding_x = None
+
+    for i in range(len(y_values) - 2):
+        y_triplet = y_values[i:i + 3]
+        average_difference = sum(abs(y - offset) for y in y_triplet) / 3
+
+        if average_difference > max_average_difference:
+            max_average_difference = average_difference
+            corresponding_x = x_values[i + 1]
+
+    return corresponding_x, max_average_difference
